draft rank
a script to run two models on the same input.
Model A (main) picks next token greedily. For model 2 (draft candidate)
we find the rank of the correct token in the logits to estimate how good
that model would be as a draft model for the specific test prompt.

For example,

```
A - Llama3-70B-Q8
B - Llama3-8B-Q5_K_M

0: 612
1: 53
2: 16
3: 4
4: 2
7: 1
15: 1

A - Llama3-70B-Q8
B - Llama3-8B-Q3_K_M

0: 604
1: 52
2: 21
3: 6
4: 2
5: 2
10: 1
16: 1

```

Note that in current linear implementation the only thing that matters
is 'is rank 0 or not'. With tree based lookup/multiple sequences it
changes.